Update to the `esbuild` v0.17 API.
I looked at using `esbuild`'s new `serve` functionality (https://esbuild.github.io/api/#serve), but that requires serving from the *output* directory and our API is based on the JS living in the *input* directory with the static files. We cannot safely let `esbuild` write into the input directory, so we maintain our own custom server for now.
If/when `esbuild` lands built-in globbing support, I might look into this again, as it might be worth finding a way to go all-in on `esbuild`'s API.